v2.4.6

+++ src/plus/register.py
@@ -26,6 +26,7 @@
 import sys
 import os
 import shelve
+import dbm
 import portalocker
 from pathlib import Path
 
@@ -36,12 +37,15 @@ from plus import config, util
 register_semaphore = QSemaphore(1)
 
 uuid_cache_path = util.getDirectory(config.uuid_cache,share=True)
+uuid_cache_path_lock = util.getDirectory(config.uuid_cache + "_lock",share=True)
 
 
 def addPathShelve(uuid,path,fh):
+    config.logger.debug("register: addPathShelve(%s,%s,_fh_)",uuid,path)
     try:
         with shelve.open(uuid_cache_path) as db:
-            db[uuid] = str(path)         
+            db[uuid] = str(path)
+        config.logger.debug("register: DB type: %s", str(dbm.whichdb(uuid_cache_path)))
     except Exception as e:
         _, _, exc_tb = sys.exc_info()
         config.logger.error("register: Exception in addPathShelve(%s,%s) line: %s shelve.open (1) %s",str(uuid),str(path),exc_tb.tb_lineno,e)
@@ -51,10 +55,13 @@ def addPathShelve(uuid,path,fh):
             config.logger.info("register: clean uuid cache %s",str(uuid_cache_path))
             # note that this deletes all "uuid" files including those for the Viewer and other files of that name prefix like uuid.db.org!!
             for p in Path(Path(uuid_cache_path).parent).glob("{}*".format(config.uuid_cache)):
-                p.unlink()
+                if str(p) != uuid_cache_path_lock:
+                    # if not the lock file, delete:
+                    p.unlink()
             # try again to acccess/create the shelve file
             with shelve.open(uuid_cache_path) as db:
                 db[uuid] = str(path)
+            config.logger.info("register: Generated db type: %s", str(dbm.whichdb(uuid_cache_path)))
         except Exception as e:
             config.logger.error("register: Exception in addPathShelve(%s,%s) line: %s shelve.open (2) %s",str(uuid),str(path),exc_tb.tb_lineno,e)                    
     finally:
@@ -63,10 +70,11 @@ def addPathShelve(uuid,path,fh):
 
 # register the path for the given uuid, assuming it points to the .alog profile containing that uuid
 def addPath(uuid,path):
+    config.logger.debug("register: addPath(%s,%s)",uuid,path)
     try:
-        config.logger.debug("register:addPath(%s,%s)",str(uuid),str(path))
         register_semaphore.acquire(1)
-        with portalocker.Lock(uuid_cache_path, timeout=0.5) as fh:
+        config.logger.debug("register:addPath(%s,%s)",str(uuid),str(path))
+        with portalocker.Lock(uuid_cache_path_lock, timeout=0.5) as fh:
             addPathShelve(uuid,path,fh)
     except portalocker.exceptions.LockException as e:
         _, _, exc_tb = sys.exc_info()
@@ -75,9 +83,9 @@ def addPath(uuid,path):
         # we remove the lock file and retry with a shorter timeout
         try:
             config.logger.info("register: clean lock %s",str(uuid_cache_path))
-            Path(uuid_cache_path).unlink()  # the lock file is not called *_lock for the uuid register as for the sync and account locks!
+            Path(uuid_cache_path_lock).unlink()
             config.logger.debug("retry register:addPath(%s,%s)",str(uuid),str(path))
-            with portalocker.Lock(uuid_cache_path, timeout=0.3) as fh:
+            with portalocker.Lock(uuid_cache_path_lock, timeout=0.3) as fh:
                 addPathShelve(uuid,path,fh)
         except portalocker.exceptions.LockException as e:
             _, _, exc_tb = sys.exc_info()
@@ -94,14 +102,17 @@ def addPath(uuid,path):
     
 # returns None if given uuid is not registered, otherwise the registered path
 def getPath(uuid):
+    config.logger.debug("register: getPath(%s)",uuid)
     try:
-        config.logger.debug("register:getPath(%s)",str(uuid))
         register_semaphore.acquire(1)
-        with portalocker.Lock(uuid_cache_path, timeout=0.5) as fh:
+        config.logger.debug("register:getPath(%s)",str(uuid))
+        with portalocker.Lock(uuid_cache_path_lock, timeout=0.5) as fh:
             try:
                 with shelve.open(uuid_cache_path) as db:
                     try:
-                        return str(db[uuid])
+                        res_path = str(db[uuid])
+                        config.logger.debug("register:getPath(%s): %s",str(uuid),res_path)
+                        return res_path
                     except:
                         return None
             except Exception as e:
@@ -117,14 +128,16 @@ def getPath(uuid):
         # we couldn't fetch this lock. It seems to be blocked forever (from a crash?)
         # we remove the lock file and retry with a shorter timeout
         try:
-            config.logger.info("register: clean lock %s",str(uuid_cache_path))
-            Path(uuid_cache_path).unlink()
+            config.logger.info("register: clean lock %s",str(uuid_cache_path_lock))
+            Path(uuid_cache_path_lock).unlink()
             config.logger.debug("retry register:getPath(%s)",str(uuid))
-            with portalocker.Lock(uuid_cache_path, timeout=0.3) as fh:
+            with portalocker.Lock(uuid_cache_path_lock, timeout=0.3) as fh:
                 try:
                     with shelve.open(uuid_cache_path) as db:
                         try:
-                            return str(db[uuid])
+                            res_path = str(db[uuid])    
+                            config.logger.debug("register:getPath(%s): %s",str(uuid),res_path)
+                            return res_path
                         except:
                             return None
                 except Exception as e:
@@ -153,6 +166,7 @@ def getPath(uuid):
 
 # scanns all .alog files for uuids and registers them in the cache
 def scanDir(path=None):
+    config.logger.debug("register: scanDir(%s)",path)
     try:
         config.logger.debug("register:scanDir(" + str(path) + ")")
         if path is None:
